Stuff that makes stuff work i dont wanna write a proper message rn im sorry

+++ src/commands/purge.command.js
@@ -3,6 +3,8 @@ const main = require("../index");
 const discord = main.discord;
 const logger = main.logger;
 
+const { Permissions } = require("discord.js");
+
 /**
  * 
  * Command that lets a user delete a certain amount of messsages from a channel.
@@ -12,14 +14,23 @@ const logger = main.logger;
  * 
  */
 async function execute(interaction) {
+    const serverId = interaction.guild_id;
+    const memberId = interaction.member.user.id;
     const channelID = interaction.channel_id;
+
+    const server = await discord.guilds.fetch(serverId);
+    const member = await server.members.fetch(memberId);
     const channel = await discord.channels.fetch(channelID);
 
-    let amount = interaction.data.options ? interaction.data.options[0].value : 100;
-    let deletedMessages = await channel.bulkDelete(amount, { filterOld: true });
-    let deleted = deletedMessages.size;
-    util.sendMessage(`Removed ${deleted} messages!`, interaction);
+    if (util.hasPermission(member, Permissions.FLAGS.MANAGE_MESSAGES)) {
+        let amount = interaction.data.options ? interaction.data.options[0].value : 100;
+        let deletedMessages = await channel.bulkDelete(amount, { filterOld: true });
+        let deleted = deletedMessages.size;
+        util.sendMessage(`Removed ${deleted} messages!`, interaction);
 
+    } else {
+        util.sendMessage("You do not have permission to use this command!", interaction);
+    }
 }
 
 const data = {
@@ -37,4 +48,4 @@ const data = {
     ],
 };
 
-module.exports = { data, execute };
+module.exports = { execute: execute, data: data, enabled: true };
